Dashboard cleanup Moved compute dimensions to its own folder to enhance readability of the component
